+++ src/sentry/static/sentry/app/views/eventsV2/table/columnEditCollection.tsx
@@ -13,10 +13,10 @@ import {t} from 'app/locale';
 import {SelectValue, OrganizationSummary} from 'app/types';
 import space from 'app/styles/space';
 import theme from 'app/utils/theme';
-import {Column, AGGREGATIONS, FIELDS, TRACING_FIELDS} from 'app/utils/discover/fields';
+import {Column} from 'app/utils/discover/fields';
 
-import {FieldValue, FieldValueKind} from './types';
-import {ColumnEditRow} from './columnEditRow';
+import {FieldValue } from './types';
+import {generateFieldOptions} from 'app/views/eventsV2/utils.tsx';
 
 type Props = {
   // Input columns
@@ -52,7 +52,7 @@ class ColumnEditCollection extends React.Component<Props, State> {
     draggingTargetIndex: void 0,
     left: void 0,
     top: void 0,
-    fieldOptions: this.generateFieldOptions(),
+    fieldOptions: generateFieldOptions(this.props.organization, this.props.tagKeys),
   };
 
   componentDidMount() {
@@ -87,71 +87,8 @@ class ColumnEditCollection extends React.Component<Props, State> {
   portal: HTMLElement | null = null;
   dragGhostRef = React.createRef<HTMLDivElement>();
 
-  generateFieldOptions() {
-    const {organization, tagKeys} = this.props;
-
-    let fields = Object.keys(FIELDS);
-    let functions = Object.keys(AGGREGATIONS);
-
-    // Strip tracing features if the org doesn't have access.
-    if (!organization.features.includes('transaction-events')) {
-      fields = fields.filter(item => !TRACING_FIELDS.includes(item));
-      functions = functions.filter(item => !TRACING_FIELDS.includes(item));
-    }
-    const fieldOptions: Record<string, SelectValue<FieldValue>> = {};
-
-    // Index items by prefixed keys as custom tags
-    // can overlap both fields and function names.
-    // Having a mapping makes finding the value objects easier
-    // later as well.
-    functions.forEach(func => {
-      const ellipsis = AGGREGATIONS[func].parameters.length ? '\u2026' : '';
-      fieldOptions[`function:${func}`] = {
-        label: `${func}(${ellipsis})`,
-        value: {
-          kind: FieldValueKind.FUNCTION,
-          meta: {
-            name: func,
-            parameters: AGGREGATIONS[func].parameters,
-          },
-        },
-      };
-    });
-
-    fields.forEach(field => {
-      fieldOptions[`field:${field}`] = {
-        label: field,
-        value: {
-          kind: FieldValueKind.FIELD,
-          meta: {
-            name: field,
-            dataType: FIELDS[field],
-          },
-        },
-      };
-    });
-
-    if (tagKeys !== null) {
-      tagKeys.forEach(tag => {
-        const tagValue =
-          FIELDS.hasOwnProperty(tag) || AGGREGATIONS.hasOwnProperty(tag)
-            ? `tags[${tag}]`
-            : tag;
-        fieldOptions[`tag:${tag}`] = {
-          label: tag,
-          value: {
-            kind: FieldValueKind.TAG,
-            meta: {name: tagValue, dataType: 'string'},
-          },
-        };
-      });
-    }
-
-    return fieldOptions;
-  }
-
   syncFields() {
-    this.setState({fieldOptions: this.generateFieldOptions()});
+    this.setState({fieldOptions: generateFieldOptions(this.props.organization, this.props.tagKeys)});
   }
 
   keyForColumn(column: Column, isGhost: boolean): string {
